Removed dependency on specific matplotlib version
We forced the use of matplotlib version 3.0.2 because newer versions of
matplotlib would raise a 'NotImplementedError' if a set_aspect('equal')
method was called.

There is no module which relies on said specific version anymore. We can therefore
relax the requirement. More on the set_aspect() issue in matplotlib:

matplotlib issue 1077

I also added 'tigl3' as an explicit dependency (again). I think all
packages which are explicitly imported/used should be listed in
``environment.yml``, otherwise it becomes hard to track what is required
and what is not. See also the note I made in the ``environment.yml`` file.
Feel free to comment on this, if you disagree :)